pyproject.toml
[tool.poetry]
name = "carsapi"
version = "0.1.0"
description = ""
authors = ["saequus <slava@spetsyian.com>"]
packages = [
    { include = "src" },
    { include = "src/**/*.py" }
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.20.0"
black = "^21.6b0"
isort = "^5.9.2"
pydantic = "^1.8.2"
mock = "^4.0.3"
tox = "^3.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
